backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# ── ML / Data ─────────────────────────────────────────────────────────────────
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
xgboost==2.0.3

# ── MLflow ────────────────────────────────────────────────────────────────────
mlflow==2.13.0

# ── Monitoring ────────────────────────────────────────────────────────────────
prometheus-client==0.20.0

# ── Config ────────────────────────────────────────────────────────────────────
pyyaml==6.0.1
pydantic==2.7.1
pydantic-settings==2.2.1

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
